Adding a varArr ADT to libutil.
This adds a variable length array to the utility library.  This ADT is
meant to provide a reusable interface to a auto-resizing array.  A
nitpick that needs to be thought about is how to handle array indices
when removing elements in the middle.  This work at the moment, however
all element above the removed one slide downwards, making it very hard
to know from outside where the element that was added might be after a
while.  If this is required, then it it probably sensible to add an
associative variable array as well.  This one here can than act as a
back-end.