Requiring the x-dimension to be completely local.
The x-dimension is the one that is local.  This means the correct choice
for the processor grid is '1 ny nz', where ny and nz are either set by
the user, or conjured up by MPI_Dims_create.  In any case, ny*nz = Np,
with Np being the total number of MPI tasks.