Keeping track of the normalisation of FFTs.
Adds functionality to retrieve the total size of a grid (ie. dim0 * dim1
* dim2) from the grid.  This can then be used to compute the
normalisation parameter of the FFT, which is for FFTW3 the inverse.
(The normalisation parameter is the number an array needs to be
multiplied with to be unchanged after tranformed forward and then
backward again [or vice versa]).