Added a tool to generate root Silo file.
This is pretty alpha at the moment, the only thing it does so far, is to
write the meshes from a Silo file (with domains and multiple files) into
a neat multimesh.  It is also quite ugly and needs major rehauling
before it is working nicely, but it serves as the proof-of-concept.