Adjusting the qay the patch handles its data.
The patch now does not allocate data arrays unless explicitly requested
(gridPatch_allocateVarData) or implicitly required
(gridPatch_getVarDataHandle).  Also a function to free the data is
provided which allowed to throw away the original arrays in the FFTs,
reducing the memory requirement there to twice the array to be
transformed instead of three times.