Added reading routines for liblare.
Currently an untested version of the legacy format (as used before by
the guys, which is somehow a paste from BDM with some Voodoo to get the
Lagrangian Coordinates).  Apparently the first few lines may be
different depending on the version, so there is some work needed to make
it work for all possible versions.