The grafic reader now autodetects whitenoise.
When a new grafic ADT is generated from a file (grafic_newFromFile()) it
will now figure out by itself whether it is a white noise file or not,
the boolean flag for white noise is hence not needed anymore.

The grafic2bov converter now uses the grafic and bov ADTs.