Migrates configure to allow for inc/lib dirs.
This renames the --with-sprng-dir option to --with-sprng-prefix and adds
--with-sprng-inc-dir and --with-sprng-lib-dir to allow to directly
specify the include and library directories for SPRNG.  This change is
driven by compiling ginnungagap on curie, where the loaded modules are
not directly accessible through LDFLAGS and CPPFLAGS but by making
macros with the include and library directory available.

Further changes will migrate the other used libraries to this system as
well.