Adds --verfiy and --initOnly to ginnungagap.
Both can be used to stop the code after parsing the ini file and after
the initialisation phase, respectively.  Both can be executed without
allocating memory for the grids, ie. the modes could be used on the
login node of a cluster to check that everything is as expected and
before submitting the job to the queue.