Adds option for predefined systems to configure.
This now allows to use --system=SYSTEM with the configure script to
configure the code according to a predefined set for a given machine.
No actual predefined system is included yet, only the system 'full' is
available, which will switch on all features the code has.

The users may define their own systems by means of writing their own
definition in an extra file (configure.systems.own) that will be
evaluated if it exists.  An example of this file is included in
doc/examples/.