Changes how MPI is handled in the build system.
This will now explicitly set and use MPICC.  Makefiles that deal with
code that has to be compiled with MPI now will set CC to MPICC.  That
allows to compile code that has no dependency on MPI without any
reference to MPI.  Implementation of this change required adjusting most
of the Makefiles and additionally a few irrelevant MPI dependencies have
been removed from a few test functions.

The way MPI is handled in by the configure is now:  --with-mpi will take
an argument that is either no to not use MPI at all, or yes, which will
use MPI and assume the MPI compiler to be 'mpicc', or any other word
that will then be taken as the value for MPICC.  This should now be the
executable without any path.  If the MPICC is in a non-standard
directory, then --with-mpi-bin-dir can be used to point to the proper
directory in which MPICC can be found.

The configure script will now report which MPICC is being used and
include the full path in the output.  The path is either taken from the
user via --with-mpi-bin-dir or deduced via 'which MPICC' (that requires
MPICC to be in the PATH).

All these changes are now used in the example configure.systems.own (in
doc/devel/) to allow for a compilation without sourcing any module.
Note that the build system does not try to resolve runtime dependency,
i.e. to run the code the user needs to ensure that all shared libraries
are actually found by the linker.