Merges changes to allow gas ICs to be generated.
This includes changes to grafic2gadget to allow the generation of gas
particles.  The IDs of the gas particles will be arrange in such a way
that they range after the DM IDs, this should keep the IDs for DM
particles identical between (full-box) runs with and without gas.  Gas
particles will have the same velocity as DM particles but are shifted by
a constant fraction of a cell.