Generating a histogram of the white noise.
If histograms are enabled, this will now also produce a histogram for
the white noise data.  This requires two additional new fields in the
ini file to set the extreme values for the white noise and the name of
the file the white noise histogram is written to.  This is analogous to
the current handling of histograms.