Adjusts validation of input parameters.
The largest sigma that is expected to occur in the white noise field is
now calculated and used to estimate whether the starting redshift is
hight enough.  The starting redshift is high enough, if the largest
overdensity is less than one.  The largest overdensity is given by the
density variance in the box (sigmaBox) times the largest expected value
in the white noise field.  The code will now terminate, if the largest
(expected) overdensity is larger than 1.