Major HDF5 update: supporting read/write.
This merges the HDF5 writing and reading for ginnungagap.  Note that
while writing basically supports compression and checksumming, the
reader does not.  Also, the reader currently only works for white noise
fields generated by realSpaceConstraints.  This limitation is due to the
ambiguity of how to map file datasets to memory datasets.