Overhauling the grid writer.
The writer now uses a streamlined code structure under the hood and
allows now for more flexibility in writing.  The construction of writers
from ini files has been factored out into a factory module and during
that, the structure of the ini files changed slightly.

The grafic writer now will not cycle through the _delta, _velx and so on
names anymore but relies on the user to set the proper file name.  The
HDF5 writer properly deals with multiple calls.  The Silo writer has
been adjusted to reflect the changes in the code structure.

As a main feature, the grid writer now will never overwrite a file
unless explicitly requested by the user.

On a side, the whole code basis has been adjusted to now work with all
three grid writers for generating velocity fields et al.  This required
including explicit renaming calls in ginnungagap to keep the names in
the files (when the file type supports that: HDF5, SILO) consistent with
what is actually in the file.  Also, ginnungagap will now write separate
files for the overdensity and for each of the velocity components
(possibly spilt over many subfiles, as in the case of Silo in MPI
mode).