Updates the grid reader infrastructure.
The grid reader now uses a similar setup as the grid writer, common IO
functionality between reading and writing is outsourced to a separate
file.  This changes also changes the INI format for the grid reader,
most notably, the type is now specified as 'type = ..' instead of
'readerType = ..'.  Additionally, all reader will now read a filename
split up into path, prefix, qualifier, and suffix; this replaces the old
'graficFileName = ..' et al.

Furthermore, the grid reading is now included in the unit testing of the
grid library (at least for BOV and HDF5) and the code documentation has
been updated.  The sample ini file for 'realSpaceConstraints' has been
adjusted to work with the new reading routines.

The main code (and tools) only had to be changed trivially to account
for the change.