[FIX] Resolves the permutation error in OpenMP.
This solves the problem that the velocity fields in the OpenMP only run
were permuted.  This was done by finally evaluating the actual
permutation of the grid instead of relying on a hardcoded permutation in
g9pInit.c that was only correct for the MPI case.  The grid now has an
additional function that returns the current dimenions (or rather, the
index with 0 being x, 1, being y, and so on) of an original dimension.