Adds functionality for doing mask IO from/to files
This allows to create an (empty) grid structure from a given mask. The
grid will have patches according to the tiling of the mask. Moving the
originally local function in the IO part to the main mask part.

Also allows getting and setting of tile data in the mask, which is used
in the reading and writing of mask data to binary files.