Adds a Mask Creator
Enables the mask creator to tag cells; the heavy-lifting function comes
for the legacay makeMask program. A test is included to verify that
tagging no cells produces a mask that is filled on the lowest level.
The mask creation not only tags the cells with their respective level,
but also runs a correction routine that will ensure that cells that are
supposed to be created on a lower grid than the mask are properly
block.

The cell count returned from the mask is now always relative to the
actual level the cells belong to.

Example: Let h be a hierarchy of grids starting at a 1^3 and going in
factors of 2 down to a 16^3 grid.  Let m be a mask on that hierarchy
with a min level of 2, a mask level of 3, and a max level of 4.  Assume
that all cells in the mask are tagged to be on the mask level, except
for a 2^3 block which on the min level and two cells that are on the max
level.
Then the number of cells reported shall be 1, 502, and 16 for level 2,
3, and 4, respectively.

A variety of tests is included:

 * A simple test (adding one high-res cell) has been added.

 * Verifying that the wrapping of the shapelet over multiple tiles works
   as expected by setting the high-res cells at the upper left back
   corner of the 0 tile. This forces to touch each tile in the mask.

 * As before, but setting the high res cells at the upper left back
   corner of the 7 tile, i.e. forcing a full periodic wrapping.

 * Tests that tagging multiple cells clustered at one place works as
   expected.

 * Test to verify that the tagging of two cells whose shapelet only
   slightly overlap works.