Adds new timer function signature.
Previously, the timer was called with timer_start("text") which would
start a timer and print something to the screen, similar for
timer_stop().  This behaviour has been moved to two new functions,
timer_start_text() and timer_stop_text(), the functions without _text
will simply start/stop the timer now.

Adjusts the legacy code to cope with this change.