Adds options to choose units in grafic2gadget.
This adds the command line options --posFactor and --velFactor to
grafic2gadget, which work as multiplier when last touching the
position/velocity arrays.  The position factor is a bit special, as it
also is used to write the boxsize parameter of the Gadget file header in
the appropriate unit.

Using
  grafic2gadget ic_velx ic_vely ic_velz icMpc.gadget
and
  grafic2gadget --posFactor 1000. ic_velx ic_vely ic_velz ickpc.gadget
will produce two Gadget files, the first one with positions in Mpc/h,
the second one with positions in kpc/h.