Adds an ID generator.
The ID generator should provide a centralized object to convert
positions to IDs and vice-versa.  IDs are fixed calculated as the
linear index on the highest resolution grid.  This assumes that for
zoomed simulations IDs are sparse.

Translating from an ID to a grid coordinate will wrap IDs, so that it is
possible to uses IDs larger than the number of cells on the finest grid
(possibly to allow for multiple particle species).  On a 2x2x2 grid IDs
0, 8, 16, 24, and so on resolve all to the (0,0,0) cell.