[FIX] Particle bunch didn't reset the data store.
When deleting a particle bunch, all actual memory referred to by
pointers stored in the data vararr is freed, however the vararr itself
must be truncated as to not trigger an assertion in the vararr deletion.

Calling truncate after freeing the actual data is safe, as the vararr
will only contain NULL pointers.