[FIX] Fixes issue with long ids in Gadget module.
When calculating the size of an element in a block, the ID block needs
to be treated separately from the floating-point blocks, i.e. the
previous check whether the file is written in double precision is not
correct to get the size of ID (which is quite independent on what the
floating point precision is).

Fixed by checking if the block for to calculate the size is the ID block
and then deducing the size per element from the state of the long ID
flag of the header code structure (note that Gadget files do not store
the width of the IDs explicitly, they have to be deduced from the block
size).

Adds a new test to check the sizes of elements in blocks to verify that
all combinations of single/double precision and 32/64 bit IDs report the
correct sizes.

Also corrects the description of the flag values in the doxygen comment
for gadgetHeader_setFlagDoublePrecision().