Adds enum to identify field types.
The field ID may be used to identify which field (velocity component,
white noise, ...) is being talked about.  Also provides a common short
name for those which can be used in string.

The datastore module now uses the field ID to figure out which field is
requested and thus how to qualify the file name.