Allows to retrieve a patch for a tile
Adds a convenience function to the mask object that allows to retrieve
an empty grid patch corresponding to a specific tile in the mask.  Also
re-factors a bit to share code between the two grid related convenience
functions in the mask.  Also adds a test for the new function,
leveraging a previous test.