[WIP] Adds HDF5 reading
The reader are generated from the ini file.  The ini file reading has
been revamped to leverage of the standard ginnungagap config (i.e. the
boxsize and zInit are read from the [Ginnungagap] section).

The whole business with the fake mask is a bit fishy, to get the cell
count work out I had to hand-tune the hierarchy to start from 1x1x1 in
factors of 2.